+++ lib/app/session_page.dart
@@ -1,0 +1,100 @@
+import 'dart:math';
+
+import 'package:confwebsite2023/core/components/responsive_widget.dart';
+import 'package:confwebsite2023/core/theme.dart';
+import 'package:confwebsite2023/features/footer/ui/footer.dart';
+import 'package:confwebsite2023/features/header/ui/header_widget.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'session_page.g.dart';
+
+@riverpod
+String sessionId(SessionIdRef ref) => throw UnimplementedError();
+
+class SessionPage extends HookConsumerWidget {
+  const SessionPage({super.key});
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      backgroundColor: baselineColorScheme.ref.secondary.secondary10,
+      body: const _MainPageBody(),
+      appBar: const LogoOnlyHeaderBar(),
+      extendBodyBehindAppBar: true,
+    );
+  }
+}
+
+class _MainPageBody extends StatelessWidget {
+  const _MainPageBody();
+
+  @override
+  Widget build(BuildContext context) {
+    final width = MediaQuery.sizeOf(context).width;
+    final largeScreenSize = ResponsiveWidget.largeScreenSize.toDouble();
+    final horizontal = max<double>(16, (width - largeScreenSize) / 4.0);
+    final padding = EdgeInsets.symmetric(
+      horizontal: horizontal,
+    );
+
+    return Stack(
+      children: [
+        const SizedBox(
+          width: double.infinity,
+          height: 800,
+          child: DecoratedBox(
+            decoration: BoxDecoration(
+              gradient: LinearGradient(
+                begin: Alignment.topCenter,
+                end: Alignment.bottomCenter,
+                colors: [
+                  Color(0xFF602678),
+                  Color(0x004B0082),
+                ],
+              ),
+            ),
+          ),
+        ),
+        CustomScrollView(
+          slivers: [
+            const SliverToBoxAdapter(
+              child: Spaces.vertical_80,
+            ),
+            _Sliver(
+              padding: padding,
+              child: Container(),
+              //child: const SponsorDetail(),
+            ),
+            const SliverToBoxAdapter(
+              child: Spaces.vertical_200,
+            ),
+            const SliverToBoxAdapter(
+              child: Footer(),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+}
+
+class _Sliver extends StatelessWidget {
+  const _Sliver({
+    required this.child,
+    required this.padding,
+  });
+
+  final Widget child;
+  final EdgeInsets padding;
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverPadding(
+      padding: padding,
+      sliver: SliverToBoxAdapter(
+        child: child,
+      ),
+    );
+  }
+}
